Fix features folders
