+++ StudentCouncilTracker.API/Controllers/CatalogUserController.cs
@@ -6,9 +6,9 @@ using StudentCouncilTracker.Application.Entities.Base.Dto;
 using StudentCouncilTracker.Application.Entities.Base.Filters;
 using StudentCouncilTracker.Application.Entities.Users.Dto;
 using StudentCouncilTracker.Application.Entities.Users.Interfaces;
-using StudentCouncilTracker.Application.Features.Users.Commands.CreateUser;
-using StudentCouncilTracker.Application.Features.Users.Commands.DeleteUser;
-using StudentCouncilTracker.Application.Features.Users.Commands.UpdateUser;
+using StudentCouncilTracker.Application.Features.Users.Commands.Create;
+using StudentCouncilTracker.Application.Features.Users.Commands.Delete;
+using StudentCouncilTracker.Application.Features.Users.Commands.Update;
 using StudentCouncilTracker.Application.Features.Users.Queries.Get;
 using StudentCouncilTracker.Application.Features.Users.Queries.GetById;
 using StudentCouncilTracker.Application.FileSavers;
