+++ StudentCouncilTracker.API/Controllers/EventActionController.cs
@@ -4,9 +4,9 @@ using StudentCouncilTracker.API.Models.ActionResults;
 using StudentCouncilTracker.Application.Entities.Base.Dto;
 using StudentCouncilTracker.Application.Entities.Base.Filters;
 using StudentCouncilTracker.Application.Entities.EventActions.Dto;
-using StudentCouncilTracker.Application.Features.EventActions.Commands.CreateEventAction;
-using StudentCouncilTracker.Application.Features.EventActions.Commands.DeleteEventAction;
-using StudentCouncilTracker.Application.Features.EventActions.Commands.UpdateEventAction;
+using StudentCouncilTracker.Application.Features.EventActions.Commands.Create;
+using StudentCouncilTracker.Application.Features.EventActions.Commands.Delete;
+using StudentCouncilTracker.Application.Features.EventActions.Commands.Update;
 using StudentCouncilTracker.Application.Features.EventActions.Queries.Get;
 using StudentCouncilTracker.Application.Features.EventActions.Queries.GetById;
 using StudentCouncilTracker.Application.OperationResults;
