+++ StudentCouncilTracker.API/Controllers/EventActionTypeController.cs
@@ -4,9 +4,9 @@ using StudentCouncilTracker.API.Models.ActionResults;
 using StudentCouncilTracker.Application.Entities.Base.Dto;
 using StudentCouncilTracker.Application.Entities.Base.Filters;
 using StudentCouncilTracker.Application.Entities.EventActionTypes.Dto;
-using StudentCouncilTracker.Application.Features.EventActionTypes.Commands.CreateEventActionType;
-using StudentCouncilTracker.Application.Features.EventActionTypes.Commands.DeleteEventActionType;
-using StudentCouncilTracker.Application.Features.EventActionTypes.Commands.UpdateEventActionType;
+using StudentCouncilTracker.Application.Features.EventActionTypes.Commands.Create;
+using StudentCouncilTracker.Application.Features.EventActionTypes.Commands.Delete;
+using StudentCouncilTracker.Application.Features.EventActionTypes.Commands.Update;
 using StudentCouncilTracker.Application.Features.EventActionTypes.Queries.Get;
 using StudentCouncilTracker.Application.Features.EventActionTypes.Queries.GetById;
 using StudentCouncilTracker.Application.OperationResults;
