+++ StudentCouncilTracker.API/Controllers/EventController.cs
@@ -4,9 +4,9 @@ using StudentCouncilTracker.API.Models.ActionResults;
 using StudentCouncilTracker.Application.Entities.Base.Dto;
 using StudentCouncilTracker.Application.Entities.Base.Filters;
 using StudentCouncilTracker.Application.Entities.Events.Dto;
-using StudentCouncilTracker.Application.Features.Events.Commands.CreateEvent;
-using StudentCouncilTracker.Application.Features.Events.Commands.DeleteEvent;
-using StudentCouncilTracker.Application.Features.Events.Commands.UpdateEvent;
+using StudentCouncilTracker.Application.Features.Events.Commands.Create;
+using StudentCouncilTracker.Application.Features.Events.Commands.Delete;
+using StudentCouncilTracker.Application.Features.Events.Commands.Update;
 using StudentCouncilTracker.Application.Features.Events.Queries.Get;
 using StudentCouncilTracker.Application.Features.Events.Queries.GetById;
 using StudentCouncilTracker.Application.OperationResults;
