+++ StudentCouncilTracker.API/Controllers/EventTypeController.cs
@@ -4,9 +4,9 @@ using StudentCouncilTracker.API.Models.ActionResults;
 using StudentCouncilTracker.Application.Entities.Base.Dto;
 using StudentCouncilTracker.Application.Entities.Base.Filters;
 using StudentCouncilTracker.Application.Entities.EventTypes.Dto;
-using StudentCouncilTracker.Application.Features.EventTypes.Commands.CreateEventType;
-using StudentCouncilTracker.Application.Features.EventTypes.Commands.DeleteEventType;
-using StudentCouncilTracker.Application.Features.EventTypes.Commands.UpdateEventType;
+using StudentCouncilTracker.Application.Features.EventTypes.Commands.Create;
+using StudentCouncilTracker.Application.Features.EventTypes.Commands.Delete;
+using StudentCouncilTracker.Application.Features.EventTypes.Commands.Update;
 using StudentCouncilTracker.Application.Features.EventTypes.Queries.Get;
 using StudentCouncilTracker.Application.Features.EventTypes.Queries.GetById;
 using StudentCouncilTracker.Application.OperationResults;
